requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
lightning=2.0.9=pyhd8ed1ab_0
lightning-cloud=0.5.38=pyhd8ed1ab_0
lightning-utilities=0.9.0=pyhd8ed1ab_0
matplotlib=3.7.2=py311h06a4308_0
matplotlib-base=3.7.2=py311ha02d727_0
matplotlib-inline=0.1.6=py311h06a4308_0
numpy=1.25.2=py311h08b1b3b_0
numpy-base=1.25.2=py311hf175353_0
pip=23.2.1=py311h06a4308_0
python=3.11.5=h955ad1f_0
pytorch=2.0.1=cpu_py311h6d93b4c_0
pytorch-lightning=2.0.9=pyhd8ed1ab_0
pyyaml=6.0=py311h5eee18b_1
torchmetrics=1.1.2=pyhd8ed1ab_0
torchvision=0.15.2=py311_cpu
